Use 'wrapper' classes to simplify pyside renderer
